# api.yml
openapi: 3.0.3
info:
  title: Sports Tournament Management API
  description: API for managing sports tournaments, teams, matches, and organization committees
  version: 1.0.0
servers:
  - url: 'https://api.example.com/v1'
    description: Production server
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: SESSION
  schemas:
...
    Org:
      type: string
      enum: [CHIEF, ORG, NONE]
...
    OrgCom:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        logo:
          type: string
        role:
          $ref: '#/components/schemas/Org'
      required:
        - name
    OrgComDTO:
      type: object
      properties:
        name:
          type: string
        logo:
          type: string
      required:
        - name
    Tournament:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        org_id:
          type: string
          format: uuid
        city:
          type: string
        min_member:
          type: integer
        orgcom_id:
          type: string
          format: uuid
        register_deadline:
          type: string
          format: date-time
        logo:
          type: string
        description:
          type: string
        is_stopped:
          type: boolean
      required:
        - name
        - org_id
...
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: integer
      required:
        - message
        - code
paths:
...
  /org_com/new:
    get:
      summary: Get new organization committee page
      security:
        - cookieAuth: []
      responses:
        '200':
          description: New organization committee page
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Пользователь не авторизован"
                code: 401
  /org_com/create:
    post:
      summary: Create new organization committee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgComDTO'
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Organization committee created
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                unauthorized:
                  value:
                    message: "Пользователь не авторизован"
                    code: 401
                empty_name:
                  value:
                    message: "Название пустое"
                    code: 400
...
  /org_com/show_all:
    get:
      summary: Show all organization committees
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of organization committees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrgCom'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: "Пользователь не авторизован"
                code: 401
  /org_com/view:
    get:
      summary: View organization committee details
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Organization committee details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrgCom'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_access:
                  value:
                    message: "Нет доступа (пользователь – не член оргкомитета)"
                    code: 403
                not_found:
                  value:
                    message: "Оркомитет не найден"
                    code: 404
...
  /org_com/edit:
    get:
      summary: Get edit organization committee page
      parameters:
        - name: orgComId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Edit organization committee page
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_access:
                  value:
                    message: "Нет доступа (пользователь – не старший орг)"
                    code: 403
                not_found:
                  value:
                    message: "Оргкомитет на найден"
                    code: 404
                    
    post:
      summary: Edit organization committee
      parameters:
        - name: orgComId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrgComDTO'
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Organization committee edited
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_access:
                  value:
                    message: "Нет доступа (пользователь – не старший орг)"
                    code: 403
                not_found:
                  value:
                    message: "Оргкомитет на найден"
                    code: 404
                empty_name:
                  value:
                    message: "Название пустое"
                    code: 400
  /org_com/tournaments:
    get:
      summary: Get tournaments for organization committee
      parameters:
        - name: orgComId
          in: query
          required: true
          schema:
            type: string
            format: uuid
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tournament'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                no_access:
                  value:
                    message: "Нет доступа (пользователь – не член оргкомитета)"
                    code: 403
                not_found:
                  value:
                    message: "Оргкомитет не найден"
                    code: 404
...